refactor(layermanager): convert LayerManagerTool to TSX and shrink to mount/unmount
